API

#API

Main.PiecewisePolynomials.PiecewisePoly — Type.

Type for function piecewise polynomial.

PiecewisePoly(
    polys::Vector{Poly},
    knots::Vector{<:Real};
    issafe::Bool = false
    )

Arguments

polys – polynomials

knots – polynomials segments

issafe – if true, correctness of agruments is not cheched

Fields

polys::Vector{Poly} – function (type depends on the basis)

knots::Vector{<:Real} – support of the function

source

Main.PiecewisePolynomials.derivative — Method.

derivative(poly::PiecewisePoly, x::Real)

Returns: the value of the first order derivative of polynomial at point x.

source

Main.PiecewisePolynomials.derivative — Method.

derivative(poly::PiecewisePoly; order::Int=1)::PiecewisePoly

Returns: derivative of order order for given polynomial.

source

Main.PiecewisePolynomials.antiderivative — Method.

antiderivative(poly::PiecewisePoly)::PiecewisePoly

Returns: antiderivative for given polynomial.

source

Main.PiecewisePolynomials.integral — Method.

integral(poly::PiecewisePoly, a::Real, b::Real)

Returns: integral of given polynomial from point a to point b.

source